io-zeromq-adapter: move header to kdbio directory
